% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{artificial_multiplets}
\alias{artificial_multiplets}
\title{Construct a dataset of artificial multiplets}
\usage{
artificial_multiplets(cell.mat, celltypes, multiplet_classes, n = 100)
}
\arguments{
\item{cell.mat}{Marker gene by barcode matrix}

\item{celltypes}{A vector of cell-type annotations}

\item{multiplet_classes}{A dataframe of multiplet types}

\item{n}{Number of artificial multiplets to construct for each multiplet type. Will also include n singlets from each cell-type.}
}
\value{
A dataframe where each row is a singlet or multiplet and the columns contain genes counts. The last column called 'type' contains the barcode type.
}
\description{
Construct a dataset of artificial multiplets
}
