% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cluster_dotplot}
\alias{cluster_dotplot}
\title{Format dataframe to create a clustered dot-plot}
\usage{
cluster_dotplot(df, rows, columns, values)
}
\arguments{
\item{df}{A dataframe in long format containing data for the dotplot}

\item{rows}{Name of dataframe column containing row labels for dotplot}

\item{columns}{Name of dataframe column containing column labels for dotplot}

\item{values}{Name of dataframe column containing values for dotplot}
}
\value{
A long format dataframe with factored and leveled row and column labels
}
\description{
Format dataframe to create a clustered dot-plot
}
