% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mroc_format}
\alias{mroc_format}
\title{Calculate multivariate receiver operator sensitivities and specificities}
\usage{
mroc_format(test, pred)
}
\arguments{
\item{test}{Test dataset used to evalute the multiplet random forest. The last column must be labeled 'type' and contain the barcode type.}

\item{pred}{Prediction probabilities for each barcode.}
}
\value{
A dataframe containing the sensitivity, specificity, and AUC for each group. Also contains the micro and macro avareages of all groups. See multiROC package documentaiton for more details.
}
\description{
Calculate multivariate receiver operator sensitivities and specificities
}
