% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mroc_plot}
\alias{mroc_plot}
\title{Plot multivariate receiver operator curves}
\usage{
mroc_plot(mroc_df, c_size = 0.3, a_size = 1)
}
\arguments{
\item{mroc_df}{multiROC result dataframe obtained from mroc_format function}

\item{c_size}{linewidth for individual barcode type curves}

\item{a_size}{linewidth for average ROC curve}
}
\value{
A ggplot object graphing multiROC curves
}
\description{
Plot multivariate receiver operator curves
}
