% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiplet_pop_sig}
\alias{multiplet_pop_sig}
\title{Compute multiplet type enrichment scores and statistical significance}
\usage{
multiplet_pop_sig(pred, sample = NULL, n = 100, homotypic = T)
}
\arguments{
\item{pred}{Multiplet type prediction probabilities for each barcode}

\item{sample}{Optional sample label annotation for each barcode. Statistical significance is assessed relative to cell-type proportions from an individual sample only.}

\item{n}{Number of simulations to run. In ach simulation multiplets are randomly created from the underlying cell-type distributins.}

\item{homotypic}{True or False; whether to include homotypic multiplets when calculating enrichment and statistical significance}
}
\value{
A dataframe containing the multiplet significance results. Each row contains data for a multiplet type in a sample. The sample, multiplet type, enrichment score, p-value, and adjusted p-value are reported
}
\description{
Compute multiplet type enrichment scores and statistical significance
}
