% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiplet_rf}
\alias{multiplet_rf}
\title{Train and test a random forest classifier to predict barcode composition}
\usage{
multiplet_rf(artificial_multiplets_mat, f = 0.8, ...)
}
\arguments{
\item{artificial_multiplets_mat}{Dataframe containing artificial multiplets. Each row is a singlet or multiplet and the columns are gene counts. The last column contains the barcode type.}

\item{f}{Fraction of multiplets to use for training the random forest}

\item{...}{Arguments passed to other methods}
}
\value{
A list containing the trained random forest, barcode-type prediction probabilities for each barcode in the test dataset, the training dataset, and the test dataset
}
\description{
Train and test a random forest classifier to predict barcode composition
}
