% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiplet_types}
\alias{multiplet_types}
\title{Enumerate all possible cell-type combinations of n cells}
\usage{
multiplet_types(celltypes, n = 2, exclude = NULL)
}
\arguments{
\item{celltypes}{A vector of cell-type annotations. There must be one annotation for each barcode.}

\item{n}{Multiplet degree, i.e. the maximum number of cells in each multiplet. Combinations will be included for integers from 2 to n.}

\item{exclude}{Optional cell-type(s) to exclude when enumerating multiplet-types.}
}
\value{
A dataframe of n columns with each row containing a multiplet type. Empty columsn are filled with NA.
}
\description{
Enumerate all possible cell-type combinations of n cells
}
