% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{neighborseq}
\alias{neighborseq}
\title{Function to run the entire Neighbor-seq pipeline}
\usage{
neighborseq(
  cell.mat,
  celltypes,
  sample = NULL,
  iter = 1,
  exclude = NULL,
  multiplet.degree = 2,
  n.am = 100,
  f = 0.8,
  nsim = 100,
  do.mroc = T,
  homotypic = T
)
}
\arguments{
\item{cell.mat}{Gene by cell matrix. Performance is improved if only marker genes are kept.}

\item{celltypes}{Vector containing cell-type annotations for each barcode}

\item{sample}{Optional vector containing sample annotations for each barcode}

\item{iter}{Optional number of iterations to train the random forest}

\item{exclude}{Optional cell-type to exclude from the artificial multiplet dataset. Can be uesful when known multiplets are present.}

\item{multiplet.degree}{Maximum number of cells in a multiplet}

\item{n.am}{Number of artificial multiplets to create for each multiplet type}

\item{f}{Fraction of artificial multiplets data to use for training the random forest}

\item{nsim}{Number of simulations to run to calculate multiplet enrichment significance}

\item{do.mroc}{True or False; whether to calculate multivariate receiver operator curves}

\item{homotypic}{True or False; whether to include homotypic multiplets when assessing enrichment and significance}
}
\value{
A list containing: 1. A combined result dataframe with multiplet enrichment scores and Fisher combined p-values if iter>1, 2. a list or dataframe with multiplet enrichment scores and p-values from each iteration, 3. a list or dataframe containing barcode type prediction probabilities for all barcodes in the dataset, 4. a list or dataframe containing the multiROC analysis
}
\description{
Function to run the entire Neighbor-seq pipeline
}
