% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_interactions}
\alias{plot_interactions}
\title{Plot the cell-cell interaction network}
\usage{
plot_interactions(
  df,
  minCounts_sample = 10,
  minCounts_total = 10,
  maxP = 0.05,
  padding = 1,
  combined = F,
  layout = "stress",
  color = NULL,
  color_breaks = NULL,
  width_breaks = NULL,
  width_range = c(0.1, 1),
  min_color = "grey50",
  max_color = "black",
  legend.position = "none",
  ...
)
}
\arguments{
\item{df}{Dataframe from containing the multiplet enrichment and significance results}

\item{minCounts_sample}{Minimum number of multiplet type counts in a sample to include the plot}

\item{minCounts_total}{Minimum number of multiplet type counts in all samples to include the plot}

\item{maxP}{Maximum multiplet type p-value to include in the plot}

\item{padding}{Line padding around cell-type labels}

\item{combined}{True or False; whether the dataframe contains an ensemble Neighbor-seq result or a single run}

\item{layout}{Network graph layout; see ggraph documentation for more information}

\item{color}{A named vector of color lables for each cell-type}

\item{color_breaks}{Edge color breaks. See ggplot documentation for more details}

\item{width_breaks}{Edge width breaks. See ggplot documentation for more details}

\item{width_range}{Edge width range. See ggplot documentation for more details}

\item{min_color}{Edge color for smallest edge / multiplet type with fewest counts}

\item{max_color}{Edge color for largest edge / multiplet type with highest counts}

\item{legend.position}{Legend position; either 'top', 'bottom', 'left', 'right', 'none'}

\item{...}{Arguments passed to other methods}
}
\value{
A ggraph object containing the cell-cell interaction network
}
\description{
Plot the cell-cell interaction network
}
