% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prep_cell_mat}
\alias{prep_cell_mat}
\title{Prepare Neighbor-seq inputs: gene expression matrix, clusters, and cluster marker genes}
\usage{
prep_cell_mat(
  ge,
  celltypes = NULL,
  logfc.threshold = 1,
  min.pct = 0.25,
  max.cells = 200,
  topn = 50,
  res = 0.8,
  ...
)
}
\arguments{
\item{ge}{Gene by barcode single-cell gene expression matrix}

\item{celltypes}{Predetermined cell cluster annotations - optional}

\item{logfc.threshold}{Minimum average log fold change of a gene in a cell-type compared to all other cell-types to be tested for differential expression. Default is 1; decreaseing the threshold will find weaker signals but will increase computation time. See documentation for Seurat FindMarkers function for more details.}

\item{min.pct}{Minimum percent of cells from a cluster to express a gene for that gene to be tested for differential expression. Default is 0.25. See documentation for Seurat FindMarkers function for more details.}

\item{max.cells}{Maximum number of cells to randomly select from each cell cluster during differential gene expression testing. Default is 200.}

\item{topn}{Number of marker genes to select for each cell-type. The union of these - there may be overlap-  will be the genes returned in the output gene expression matrix. Top arker genes are selected based on their fold-change rank.}

\item{res}{Resolution parameter for Seurat clustering. Larger values will find a larger number of clusters. See Seurat FindClusters function for more details.}

\item{...}{Arguments passed to other methods}
}
\value{
A list containing the cell matrix, cell-types, and marker gene output
}
\description{
Prepare Neighbor-seq inputs: gene expression matrix, clusters, and cluster marker genes
}
