% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xgpred}
\alias{xgpred}
\title{Barcode-type prediction using a trained random forest}
\usage{
xgpred(rf, cell.mat)
}
\arguments{
\item{rf}{Random forest trained on artificial multiplets. Output of multiplet_rf}

\item{cell.mat}{Gene by cell matrix}
}
\value{
Dataframe of barcode-type prediction probabilities
}
\description{
Barcode-type prediction using a trained random forest
}
